+++ libs/gesha-core/src/targets/rust_type/definition.rs
@@ -1,5 +1,4 @@
-use crate::targets::rust_type::{DataType, DeriveAttribute, StructField, TypeHeader};
-use heck::ToUpperCamelCase;
+use crate::targets::rust_type::{DataType, DeriveAttribute, EnumVariant, StructField, TypeHeader};
 use std::fmt::{Debug, Display, Formatter};
 
 #[derive(Clone, Debug)]
@@ -123,23 +122,5 @@ impl EnumDef {
 impl From<EnumDef> for Definition {
     fn from(this: EnumDef) -> Self {
         Self::EnumDef(this)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct EnumVariant {
-    pub compatible: String,
-    pub original: String,
-    _hide_default_constructor: bool,
-}
-
-impl EnumVariant {
-    pub fn new(original: String) -> Self {
-        // TODO: replace x with Rust compatible chars if illegal chars are included.
-        EnumVariant {
-            compatible: original.to_upper_camel_case(),
-            original,
-            _hide_default_constructor: true,
-        }
     }
 }
